Uniview
In this era, students remain confused about which university is good for their desired curriculum, or if they should take a course with a respected teacher or not. Getting feedback and learning about others' experiences help students to have a better understanding before the semester starts. Students can get to know the unknown subjects and they also find out about the courses which are matched with their field of interest so that they can get the desired faculty with understanding for their particular need (learning method). Of course, perspectives may vary but, when fellow students share their feedback in a single platform, it becomes easier to understand many things and have a clear picture for future planning. This is what our website, Unireview offers; a review of universities, courses and faculty members from different prespectives.